Remove unnecessary xml declaration
It turns out that not only the DTD is not needed, but also the xml declaration, since the values ​​used are the default values. Stack works without all this xml prolog.

More details can be found in [supplementary material](https://oreillymedia.github.io/Using_SVG/extras/ch01-XML.html) to the book “Using SVG with CSS3 and HTML5”.