Improve namespace handling
For namespaces, the values in their declarations are essential. They must be unique. Suffixes are nothing more than aliases and can be renamed, for example to remove duplicate namespaces or to remove duplicate aliases to different namespaces.

The namespace `http://www.w3.org/1999/xlink` is now deprecated. Its declaration can be removed, along with its alias, from the prefix to the `href` attribute.

It is also possible to remove the declaration of any non-default namespace if its alias is not used as attribute prefixes.